change from CG query to general external price
