+++ queries/rates/constants.ts
@@ -2,3 +2,5 @@ export const CANDLES_ENDPOINT =
 	'https://api.thegraph.com/subgraphs/name/tburm/optimism-latest-rates';
 
 export const DAY_PERIOD = 24 * 3600;
+
+export const CG_BASE_API_URL = 'https://api.coingecko.com/api/v3';
