+++ queries/rates/utils.ts
@@ -2,6 +2,7 @@ import { RateUpdates, Candles, Prices } from './types';
 import Wei, { wei } from '@synthetixio/wei';
 import { RateUpdateResult } from '@synthetixio/queries/build/node/generated/exchangesSubgraphQueries';
 import { ethers } from 'ethers';
+import { Synths } from 'constants/currency';
 
 export const getMinAndMaxRate = (rates: RateUpdateResult[]): [Wei, Wei] => {
 	if (rates.length === 0) return [wei(0), wei(0)];
@@ -51,4 +52,42 @@ export const mapLaggedDailyPrices = (prices: Candles): Prices => {
 			price: Number(candle.average),
 		};
 	});
+};
+
+const markets = [
+	Synths.sETH,
+	Synths.sBTC,
+	Synths.sLINK,
+	Synths.sSOL,
+	Synths.sAVAX,
+	Synths.sMATIC,
+	Synths.sAAVE,
+	Synths.sUNI,
+	Synths.sEUR,
+	'sXAU',
+	'sXAG',
+	'sWTI',
+] as const;
+
+const map: Record<typeof markets[number], string> = {
+	[Synths.sETH]: 'ethereum',
+	[Synths.sBTC]: 'bitcoin',
+	[Synths.sLINK]: 'chainlink',
+	[Synths.sSOL]: 'solana',
+	[Synths.sAVAX]: 'avalanche-2',
+	[Synths.sMATIC]: 'matic-network',
+	[Synths.sAAVE]: 'aave',
+	[Synths.sUNI]: 'uniswap',
+	[Synths.sEUR]: 'euro',
+	sXAU: '',
+	sXAG: '',
+	sWTI: '',
+};
+
+export const synthToCoingeckoPriceId = (synth: any) => {
+	if (markets.includes(synth)) {
+		return map[synth as typeof markets[number]];
+	} else {
+		return 'ethereum';
+	}
 };
